### 2025-01-06 #7631 @editorialbot
[PRE REVIEW]: textAnnotatoR: An Interactive Text Annotation Tool with 'shiny' GUI
**Submitting author:** @chaoliu-cl (<a href="http://orcid.org/0000-0002-9979-8272">Chao Liu</a>)
**Repository:** https://github.com/chaoliu-cl/textAnnotatoR
**Branch with paper.md** (empty if default branch): master
**Version:** v0.1.4
