### 2025-03-24 f6308b9
correct image path
