### 2025-04-01 chaoliu-cl/textAnnotatoR#1 @cmaimone
JOSS paper
Re: https://github.com/openjournals/joss-reviews/issues/7971 - JOSS paper

Please conduct a more comprehensive review of comparable packages/alternatives - not just the main commercial ones.  See https://joss.theoj.org/papers/search?q=qualitative&search_button= as a first pass
