### 2025-04-01 chaoliu-cl/textAnnotatoR#2 @cmaimone
Line breaks not included
I uploaded a text file to try the tool, and the text it showed me had the line breaks removed.  This makes the text unreadable for actually working with it in the tool.

I'm working on a Mac -- in case you have some system-specific line-ending issue going on.

Re: https://github.com/openjournals/joss-reviews/issues/7971
