### 2025-04-01 chaoliu-cl/textAnnotatoR#3 @cmaimone
choosing project storage directory
When I first run the app, I get a pop-up: "textAnnotatoR needs to create a directory at: /Users/christina/Library/Application Support/org.R-project.R/R/textAnnotatoR to store your annotation projects. Do you approve?"

Could the user have a choice over where this data is stored?  Could this be per-project?

I'm concerned about two things:
1. If I'm working with files that might have sensitive data, I might want to make sure that the project data/codes are stored in a specific location, and
2. If someone is using a employer-managed computer, or working on a shared system, whether they will necessarily have access/permissions to the default folder.

More generally, it would be helpful to have more info in the documentation on what is stored for each project.

Re: https://github.com/openjournals/joss-reviews/issues/7971
