### 2025-04-01 chaoliu-cl/textAnnotatoR#4 @cmaimone
Can't apply code
I started the app, uploaded a file, and then highlighted some text.  I opened the Code/Memo tab, and entered a value for the "Code" field.  When I click "Apply Code" button, I get an error message in the bottom right of the screen that says to select text before applying a code, but the text is still selected/highlighted.

Re: https://github.com/openjournals/joss-reviews/issues/7971
