### 2025-04-01 chaoliu-cl/textAnnotatoR#5 @cmaimone
Documentation missing: annotate_text
Documentation page https://liu-chao.site/textAnnotatoR/articles/textAnnotatoR_gui.html references an annotate_text function that is not documented -- it doesn't appear in https://liu-chao.site/textAnnotatoR/reference/index.html

If people can interact with the application through code, and not just through the GUI, that needs to be documented.

Re: https://github.com/openjournals/joss-reviews/issues/7971
