### 2025-04-01 chaoliu-cl/textAnnotatoR#5 @cmaimone
There are additional functions referenced on that page as well with examples given using them that are not documented
