### 2025-04-01 chaoliu-cl/textAnnotatoR#6 @cmaimone
Themes
![Image](https://github.com/user-attachments/assets/8091b739-2cb7-4106-b94e-a1e2c73d7079)

When I go to the theme tab (first time), I haven't created any theme, but the summary output is showing that there's one theme.  I'm assuming that one theme is maybe "Root - Root of the code hierarchy" but that wasn't clear.  That box is labeled "Code Hierarchy" not "Themes".  It's unclear if codes and themes are distinct -- I think they are, but then something is mislabeled.

Re: https://github.com/openjournals/joss-reviews/issues/7971
