### 2025-04-01 chaoliu-cl/textAnnotatoR#7 @cmaimone
Example usage
The current documentation includes high-level instructions for what you can do with the tool, but it does not provide an example use case.  It would be helpful for the documentation for the package to include an example project with real output examples.  If it's not possible to include a starter "project" directly in the package, could one be dowloaded?  So that people can replicate the steps in the tutorial if they want?

Re: https://github.com/openjournals/joss-reviews/issues/7971
