### 2025-04-01 chaoliu-cl/textAnnotatoR#8 @cmaimone
Missing vignettes?
The readme.md file mentions: "Key vignettes include: - Getting Started with textAnnotatoR - Managing Code Hierarchies - Analyzing Coding Patterns - Comparing Multiple Coders" but I only see one vignette at https://liu-chao.site/textAnnotatoR/ -- are there some missing?

Re: https://github.com/openjournals/joss-reviews/issues/7971
