### 2025-04-02 #7971 @ajstewartlang
> [@ajstewartlang](https://github.com/ajstewartlang) - I'm not sure if this package is ready for review. I wasn't able to use the basic functionality on first use, and there appear to be multiple functions that lack documentation. I've added some initial issues to the repo. I'll pause on continuing the review until those are resolved.

Many thanks for highlighting this @cmaimone.

:wave: @chaoliu-cl could you address these issues please?
