### 2025-04-04 0bbd491
updating functions and documentation
