+++ paper/paper.bib
@@ -73,3 +73,47 @@
     note = {R package version 0.9.1},
     url = {https://CRAN.R-project.org/package=readtext},
   }
+
+@article{Bogachenkova2023,
+  author = {Bogachenkova, Anastasia and others},
+  title = {LaMa: a thematic labelling web application},
+  journal = {Journal of Open Source Software},
+  year = {2023},
+  volume = {8},
+  number = {85},
+  pages = {5135},
+  doi = {10.21105/joss.05135}
+}
+
+@article{Gjerde2019,
+  author = {Gjerde, Vibeke},
+  title = {corporaexplorer: An R package for dynamic exploration of text collections},
+  journal = {Journal of Open Source Software},
+  year = {2019},
+  volume = {4},
+  number = {38},
+  pages = {1342},
+  doi = {10.21105/joss.01342}
+}
+
+@article{Proctor2024,
+  author = {Proctor, C.},
+  title = {A tool for qualitative data analysis designed to support computational thinking},
+  journal = {Journal of Open Source Software},
+  year = {2024},
+  volume = {9},
+  number = {102},
+  pages = {7031},
+  doi = {10.21105/joss.07031}
+}
+
+@article{Rampin2021,
+  author = {Rampin, Rémi and others},
+  title = {Taguette: open-source qualitative data analysis},
+  journal = {Journal of Open Source Software},
+  year = {2021},
+  volume = {6},
+  number = {68},
+  pages = {3522},
+  doi = {10.21105/joss.03522}
+}
