+++ R/annotate_gui.R
@@ -174,6 +174,8 @@ annotate_gui <- function() {
             margin-right: 50px;
           }
           #text_display {
+            white-space: pre-wrap;  /* This preserves line breaks in the display */
+            word-wrap: break-word;  /* Allows text to break at word boundaries */
             background-color: white;
             border: 1px solid #dee2e6;
             border-radius: 5px;
@@ -215,21 +217,24 @@ annotate_gui <- function() {
             height: calc(100% - 40px);
             overflow-y: auto;
           }
-          .theme-item, .code-item {
-            display: inline-block;
-            padding: 2px 4px;
-            border-radius: 3px;
-            transition: background-color 0.2s;
+          .root-item {
+            font-weight: bold;
+            color: #333;
           }
-
-          .theme-item:hover, .code-item:hover {
+          .theme-item {
+            font-weight: bold;
+            color: #0275d8;
+          }
+          .code-item {
+            color: #5cb85c;
+          }
+          .root-item:hover, .theme-item:hover, .code-item:hover {
             background-color: rgba(0, 0, 0, 0.05);
+            cursor: pointer;
           }
-
-          .theme-item.selected, .code-item.selected {
+          .root-item.selected, .theme-item.selected, .code-item.selected {
             background-color: #e6f3ff;
           }
-
           #hierarchy_pre {
             font-family: monospace;
             line-height: 1.5;
@@ -335,11 +340,11 @@ annotate_gui <- function() {
                        ),
                        column(8,
                               wellPanel(
-                                h4("Code Hierarchy"),
+                                h4("Code & Theme Hierarchy"),  # Updated label
                                 uiOutput("hierarchy_view")
                               ),
                               wellPanel(
-                                h4("Theme Details"),
+                                h4("Selected Item Details"),  # Updated label
                                 uiOutput("theme_details")
                               )
                        )
@@ -471,16 +476,68 @@ annotate_gui <- function() {
     # Initialize directory with confirmation
     observe({
       if (is.null(rv$data_dir)) {
-        init_data_dir(session)
+        rv$data_dir <- init_data_dir(session)
       }
     })
 
-    # Handle directory confirmation
-    handle_dir_confirmation(input, rv, session)
+    # Set up roots for shinyFiles
+    roots <- c(Home = path.expand("~"))
+    if (.Platform$OS.type == "windows") {
+      roots <- c(roots, getVolumes()())
+    }
+
+    # Initialize directory choosers for other parts of the app
+    # Note: The custom_dir_select is already initialized in init_data_dir
+    shinyDirChoose(input, "directory_select", roots = roots, session = session)
+    shinyDirChoose(input, "text_directory_select", roots = roots, session = session)
+    shinyDirChoose(input, "records_directory_select", roots = roots, session = session)
+    shinyFileChoose(input, "file_select", roots = roots, session = session)
+
+    # Handle selected custom directory
+    observeEvent(input$custom_dir_select, {
+      selected_dir <- parseDirPath(roots, input$custom_dir_select)
+      if (length(selected_dir) > 0 && nchar(selected_dir) > 0) {
+        session$userData$custom_dir <- selected_dir
+        # Update UI to show selected path
+        shinyjs::runjs(sprintf('document.getElementById("custom_dir_path").innerText = "%s";',
+                               gsub("\\\\", "\\\\\\\\", selected_dir)))
+      }
+    }, ignoreInit = TRUE)
+
+    # Handle final confirmation of directory
+    observeEvent(input$confirm_data_dir, {
+      if (input$storage_location == "default") {
[…]
