### 2025-04-04 #7971 @chaoliu-cl
@cmaimone Thank you very much for your comments. I have addressed the issues that you brought up. Specifically,
•	Comparable packages review: A review of comparable packages has been added to the Statement of Need section.
•	Line Breaks: Line breaks have been preserved throughout the text to ensure readability and proper formatting.
•	Directory Selection: A Choose Directory button has been added to the pop-up window, allowing users to specify where their data should be saved.
•	Clarification on Coding Process: I apologize for the confusion regarding the Apply Code button. This function is intended for applying existing codes to the text. When creating a new code, please use the Save button instead. The documentation has been updated to provide clearer instructions on this process.
•	Additional Resources: The package now includes two example files, two vignettes, and a more comprehensive documentation page to help users understand the workflow more easily.
•	Theme Tab Enhancements: Several improvements have been made to the Themes tab:
  1.	The root node's type has been changed from "theme" to "root" to distinguish it from user-defined themes.
  2.	The calculate_hierarchy_stats() function now excludes the root node from the theme count, focusing only on actual themes.
  3.	UI labels have been updated for better clarity:
    	Code Hierarchy → Code & Theme Hierarchy
    	Theme Details → Selected Item Details
  4.	Visual enhancements have been added, including distinct styling and icons for root nodes, themes, and codes.
  5.	The theme details panel now provides context-sensitive output which displays relevant information based on whether a root node, theme, or code is selected.

Please install the latest version of the package from GitHub:
`devtools::install_github("chaoliu-cl/textAnnotatoR")`
