### 2025-04-07 f4fed3f
updating functions and documentation
