+++ tools/optimize_images.R
@@ -1,0 +1,300 @@
+#' Optimize PNG images in the package
+#'
+#' @param pkg_dir Package root directory, default is current directory
+#' @param target_dirs Directories to process
+#' @param quality Quality setting for JPEG (1-100), lower means smaller files
+#' @param compression_level PNG compression level (0-9), higher means more compression
+#' @param resize Percentage to resize images (e.g., 70 means resize to 70% of original)
+#' @param backup Create backup of original images
+#' @param verbose Print detailed information
+#' @return Invisible NULL
+optimize_package_images <- function(pkg_dir = ".",
+                                    target_dirs = c("man/figures", "vignettes/figures"),
+                                    quality = 75,
+                                    compression_level = 9,
+                                    resize = 70,
+                                    backup = TRUE,
+                                    verbose = TRUE) {
+
+  # Check if required packages are available
+  if (!requireNamespace("magick", quietly = TRUE)) {
+    message("Please install the 'magick' package: install.packages('magick')")
+    return(invisible(NULL))
+  }
+
+  # Process each target directory
+  for (target_dir in target_dirs) {
+    full_dir <- file.path(pkg_dir, target_dir)
+
+    if (!dir.exists(full_dir)) {
+      if (verbose) message("Directory not found: ", full_dir)
+      next
+    }
+
+    # Get all image files in the directory
+    img_files <- list.files(full_dir,
+                            pattern = "\\.(png|jpe?g|gif)$",
+                            full.names = TRUE,
+                            ignore.case = TRUE)
+
+    if (length(img_files) == 0) {
+      if (verbose) message("No image files found in ", full_dir)
+      next
+    }
+
+    if (verbose) message("Found ", length(img_files), " image files in ", full_dir)
+
+    # Create backup directory if needed
+    if (backup) {
+      backup_dir <- file.path(full_dir, "backup")
+      if (!dir.exists(backup_dir)) {
+        dir.create(backup_dir)
+      }
+    }
+
+    # Track total size reduction
+    original_total <- 0
+    optimized_total <- 0
+
+    # Process each file
+    for (img_file in img_files) {
+      # Get file info before optimization
+      original_size <- file.info(img_file)$size
+      original_total <- original_total + original_size
+
+      # Create backup if requested
+      if (backup) {
+        backup_file <- file.path(backup_dir, basename(img_file))
+        file.copy(img_file, backup_file, overwrite = TRUE)
+      }
+
+      # Optimize the image
+      if (verbose) message("Optimizing ", basename(img_file), "...")
+
+      # Read the image
+      img <- magick::image_read(img_file)
+
+      # Get original dimensions
+      img_info <- magick::image_info(img)
+      original_width <- img_info$width
+      original_height <- img_info$height
+
+      # Resize if requested
+      if (resize < 100) {
+        new_width <- as.integer(original_width * resize / 100)
+        new_height <- as.integer(original_height * resize / 100)
+        img <- magick::image_resize(img, paste0(new_width, "x", new_height))
+
+        if (verbose) {
+          message(sprintf("  Resized from %dx%d to %dx%d",
+                          original_width, original_height,
+                          new_width, new_height))
+        }
+      }
+
+      # Adjust quality and compression based on file type
+      file_ext <- tolower(tools::file_ext(img_file))
+
+      if (file_ext %in% c("jpg", "jpeg")) {
+        img <- magick::image_write(img, img_file, format = "jpeg", quality = quality)
+      } else if (file_ext == "png") {
+        img <- magick::image_write(img, img_file, format = "png", quality = quality)
+
[…]
