+++ vignettes/technical_integration.Rmd
@@ -10,16 +10,23 @@ vignette: >
   %\VignetteIndexEntry{Technical Integration - Using textAnnotatoR with Other R Tools}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
+  %\VignetteDepends{dplyr, ggplot2}
 ---
 
 ```{r setup, include=FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  fig.width = 7,
-  fig.height = 5,
+  fig.width = 4,
+  fig.height = 3,
   fig.align = 'center',
-  fig.path = "figures/"
+  dpi = 72,
+  out.width = "60%",
+  fig.retina = NULL,
+  dev = "png",
+  fig.path = "figures/",
+  message = FALSE,
+  warning = FALSE
 )
 library(textAnnotatoR)
 ```
@@ -36,7 +43,7 @@ Before integrating with other tools, let's explore the structure of textAnnotato
 
 textAnnotatoR project files (.rds) contain the following components:
 
-```{r echo = T, results = 'hide'}
+```{r echo = TRUE, results = 'hide'}
 # First, ensure the project directory exists and is initialized
 # We need to handle this explicitly when working outside the GUI
 initialize_project_directory <- function() {
@@ -61,50 +68,27 @@ initialize_project_directory <- function() {
 # Initialize the directory
 projects_dir <- initialize_project_directory()
 
-# Now load a project with an absolute path
-project_path <- file.path(projects_dir, "my_project.rds")
-
-# If you're loading an existing project file from another location,
-# you may need to copy it to the projects directory first
-if (file.exists("C:/Users/chaoliu/Downloads/my_project.rds")) {
-  file.copy("C:/Users/chaoliu/Downloads/my_project.rds", project_path)
-}
-
-# Load the project
-if (file.exists(project_path)) {
-  project <- readRDS(project_path)
-} else {
-  # If project doesn't exist, create a sample one for demonstration
-  project <- list(
-    text = "Sample interview text content...",
-    annotations = data.frame(
-      start = c(1, 50),
-      end = c(10, 60),
-      text = c("Sample", "interview"),
-      code = c("sample_code", "another_code"),
-      memo = c("Sample memo", "Another memo"),
-      stringsAsFactors = FALSE
-    ),
-    codes = c("sample_code", "another_code"),
-    code_tree = data.tree::Node$new("Root"),
-    code_colors = c("sample_code" = "#FF0000", "another_code" = "#00FF00"),
-    memos = list(),
-    code_descriptions = list()
-  )
-  
-  # Save the sample project
-  saveRDS(project, project_path)
-  cat("Created sample project at:", project_path, "\n")
-}
+# Create a sample project for demonstration
+project <- list(
+  text = "Sample interview text content for demonstration purposes. This text contains information about remote work challenges and benefits.",
+  annotations = data.frame(
+    start = c(1, 25, 50, 80),
+    end = c(10, 35, 60, 90),
+    text = c("Sample", "interview", "demonstration", "remote work"),
+    code = c("Sample", "Interview", "Demo", "Remote"),
+    memo = c("Example memo", "Interview note", "Demo note", "Remote note"),
+    stringsAsFactors = FALSE
+  ),
+  codes = c("Sample", "Interview", "Demo", "Remote"),
+  code_tree = data.tree::Node$new("Root"),
+  code_colors = c("Sample" = "#FF0000", "Interview" = "#00FF00", 
+                  "Demo" = "#0000FF", "Remote" = "#FFCC00"),
+  memos = list(),
+  code_descriptions = list()
+)
 
 # Structure of a project
[…]
