+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: http://liu-chao.site/textAnnotatoR/
+url: ~
 template:
   bootstrap: 5
 

[…]
