+++ pkgdown.yml
@@ -2,8 +2,6 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
-  textAnnotatoR_gui: textAnnotatoR_gui.html
-last_built: 2025-03-25T01:51Z
-urls:
-  reference: http://liu-chao.site/textAnnotatoR/reference
-  article: http://liu-chao.site/textAnnotatoR/articles
+  practical_example: practical_example.html
+  technical_integration: technical_integration.html
+last_built: 2025-04-07T17:46Z
