### 2025-04-22 chaoliu-cl/textAnnotatoR#4 @cmaimone
For reference, I'm using a Mac, and I've tried both Firefox and Chrome
