### 2025-04-22 chaoliu-cl/textAnnotatoR#9 @cmaimone
No indication of what project is open/active
In the gui, there isn't an indication of whether a project is open, and if so, what that project is - so I wasn't sure if I had a project open, needed to create a new one, etc.

re: https://github.com/openjournals/joss-reviews/issues/7971
