### 2025-04-22 chaoliu-cl/textAnnotatoR#10 @cmaimone
Project won't save
I tried to save my project, and the dialog box opened, I didn't change the directory, I entered a name (tmp), and clicked save but then nothing happened

re: https://github.com/openjournals/joss-reviews/issues/7971
