### 2025-04-22 chaoliu-cl/textAnnotatoR#10 @cmaimone
I think the issue is file permissions on Macs.  If you don't click the directory button to open the file explorer and choose the location to save the file, you don't get the popup windows to prompt you to give permission to RStudio to allow you to save files.  Once I did this, I was able to save the project file.  So what's needed is probably extra instructions for Mac users related to the first time they run the application (and possibly after updates, etc. - for other programs, I sometimes have to regrant permissions)
