### 2025-04-22 #7971 @cmaimone
@chaoliu-cl - I installed the dev version, but I'm still unable to apply any codes to the text (I try to save the code first, but nothing happens when I click save).  And the line breaks still don't show up

Please update issues on your repo (and resolve them as appropriate) to help track the updates on the project.

In terms of documentation, vignettes are great, but the are not sufficient documentation.  All functions/objects that are exposed to users (that users are expected to use) should have an entry in https://liu-chao.site/textAnnotatoR/reference/index.html   But I may be misunderstanding how the package is supposed to work.  Are the functions in https://liu-chao.site/textAnnotatoR/articles/technical_integration.html examples of functions that the user could write, or are they functions included in the package (either backend functions or functions people are supposed to use)?  I would expect at least some documentation about an object as complex as the "project".
