### 2025-04-22 #7971 @chaoliu-cl
Hi @cmaimone , Thank you once again for bringing these issues to my attention. I will work on enhancing the documentation to ensure it is more comprehensive.In the meantime, I was unable to reproduce the "Apply Code" issue you mentioned. I tested it on a MacBook (please see the attached GIF), and the function appears to be working correctly. Can you try uninstalling the package using the command `remove.packages("textAnnotatoR")` and then reinstalling the development version with `devtools::install_github("chaoliu-cl/textAnnotatoR")`?

![Image](https://github.com/user-attachments/assets/2809fdd1-14eb-4ec3-8e92-3bff03034b35)
