+++ vignettes/technical_integration.Rmd
@@ -33,7 +33,9 @@ library(textAnnotatoR)
 
 # Introduction
 
-While textAnnotatoR provides a comprehensive GUI for annotation, its true power comes from integration with the broader R ecosystem. This vignette demonstrates how to integrate textAnnotatoR's outputs with other popular R packages for text analysis, visualization, and reporting.
+While textAnnotatoR provides a comprehensive GUI for annotation, it can be integrated seamlessly with the broader R ecosystem. This vignette demonstrates how you can integrate textAnnotatoR's outputs with other popular R packages for text analysis, visualization, and reporting.
+
+> **IMPORTANT NOTE:** The example functions shown in this vignette are not part of the textAnnotatoR package. They are provided as examples that users can adapt and implement in their own workflows to extend textAnnotatoR's capabilities. These sample functions demonstrate how to work with textAnnotatoR's data structures to integrate with other R tools.
 
 # Understanding Data Structures
 
@@ -44,8 +46,9 @@ Before integrating with other tools, let's explore the structure of textAnnotato
 textAnnotatoR project files (.rds) contain the following components:
 
 ```{r echo = TRUE, results = 'hide'}
-# First, ensure the project directory exists and is initialized
-# We need to handle this explicitly when working outside the GUI
+# EXAMPLE USER FUNCTION - NOT PART OF THE PACKAGE
+# This function demonstrates how a user might initialize a directory
+# for working with textAnnotatoR project files programmatically
 initialize_project_directory <- function() {
   # Get user documents directory as a safer default
   user_docs <- file.path(path.expand("~"), "Documents")
@@ -106,7 +109,7 @@ This structure makes it readily compatible with standard R data manipulation too
 
 ## Using textAnnotatoR with quanteda
 
-The [quanteda](https://quanteda.io/) package is a powerful framework for text analysis. Here's how to combine textAnnotatoR annotations with quanteda:
+The [quanteda](https://quanteda.io/) package is a powerful framework for text analysis. Here's how you can combine textAnnotatoR annotations with quanteda:
 
 ```{r eval=TRUE}
 library(quanteda)
@@ -134,7 +137,7 @@ The [tidytext](https://www.tidytextmining.com/) package works beautifully with t
 library(tidytext)
 library(ggplot2)
 
-# Convert annotations to tidy format
+# Example of how you can convert annotations to tidy format
 tidy_annotations <- project$annotations %>%
   unnest_tokens(word, text) %>%
   anti_join(stop_words)
@@ -159,12 +162,13 @@ ggplot(word_frequencies, aes(reorder(word, n), n, fill = code)) +
 
 ## Network Visualization with igraph
 
-The following example demonstrates how to create network visualizations from co-occurrence patterns:
+The following example demonstrates how you can create network visualizations from co-occurrence patterns:
 
 ```{r, eval=FALSE}
 library(igraph)
 library(ggplot2)
 
+# EXAMPLE USER CODE - NOT PART OF THE PACKAGE
 # Create a simple matrix of code co-occurrences
 codes <- c("Home Setup", "Work-Life", "Tools", "Benefits")
 co_matrix <- matrix(c(
@@ -201,6 +205,7 @@ knitr::include_graphics("../man/figures/code_cooccurance.png")
 Create static visualizations of your code frequencies:
 
 ```{r, eval=FALSE}
+# EXAMPLE USER CODE - NOT PART OF THE PACKAGE
 # Create a base ggplot with code frequencies
 code_freq <- data.frame(
   code = codes,
@@ -223,10 +228,11 @@ knitr::include_graphics("../man/figures/example_code_freq.png")
 
 ## Processing Multiple Documents
 
-For projects with multiple documents, you can automate the workflow:
[…]
