+++ pkgdown.yml
@@ -4,4 +4,4 @@ pkgdown_sha: ~
 articles:
   practical_example: practical_example.html
   technical_integration: technical_integration.html
-last_built: 2025-04-07T17:46Z
+last_built: 2025-04-23T14:48Z
