### 2025-04-23 #7971 @chaoliu-cl
Hi @cmaimone,

I've made further improvements to the documentation, and it should now be more complete. I've added explanations for the buttons related to project handling, and the vignette page with the practical example now includes a more detailed walkthrough of a sample project.

Regarding your question about the functions on the technical integration vignette (https://liu-chao.site/textAnnotatoR/articles/technical_integration.html): those functions are not part of the textAnnotatoR package itself. They are provided as examples for users to adapt when integrating the tool into their workflows. I've added a prominent note at the top of the page to clarify this, and included comments like "EXAMPLE USER FUNCTION – NOT PART OF THE PACKAGE" above each relevant function definition for added clarity.

As for your question about listing the entries of the package functions: these are auto-generated by pkgdown. Since textAnnotatoR is primarily a point-and-click GUI application, I don't believe listing internal function entries would be particularly helpful—or even feasible in some cases.

P.S. I've tested the GUI buttons and functionalities on several computers across different operating systems, and everything appears to be working correctly. If you're still experiencing issues, please try a fresh install of the package on another computer.
