### 2025-05-14 #7971 @cmaimone
@chaoliu-cl - I still can't apply codes.  The line breaks do appear now in the files though.  I recorded my screen while trying to use the app ([file here](https://nuwildcat.sharepoint.com/:v:/s/NUIT-RCS/EduvimmYS0dKpIQp84n_l00BtWV_UOEEHUyyOHR44-Q9vg?e=OOuguP)) if you want to see if I'm doing something wrong
