+++ vignettes/practical_example.Rmd
@@ -151,7 +151,7 @@ Now let's start coding our interview text:
 5. Add a memo: "Initial challenge in transition to remote work"
 6. Click "Save" to save the code
 
-<img src="C:/Users/chaoliu/OneDrive - Cedarville University/Documents/textAnnotatoR/man/figures/Save_code_gif.gif" width="700" />
+<img src="../man/figures/Save_code_gif.gif" width="700" />
 
 Continue this process for other notable segments in the text. For example:
 
@@ -173,7 +173,7 @@ Once you've created some codes, you can apply them to new text segments:
 2. Click on "Apply Code" in the side panel 
 3. Choose the code you want to apply from the dropdown menu
 
-<img src="C:/Users/chaoliu/OneDrive - Cedarville University/Documents/textAnnotatoR/man/figures/Apply_code_gif.gif" width="700" />
+<img src="../man/figures/Apply_code_gif.gif" width="700" />
 
 ### Managing Codes
 
