### 2025-05-16 #7971 @chaoliu-cl
Hi @cmaimone , thank you for the video. To highlight the text, you need to first click on the selection tool (the cursor icon) located in the right margin:

![Image](https://github.com/user-attachments/assets/7ed5336a-5474-4e78-952c-fff0ca3a3e84)

And to apply code, highlight the text using the selection tool. Then click on Apply Code, a drop-down menu will appear with a list of available codes for you to choose from.

![Image](https://github.com/user-attachments/assets/274e93a1-1a36-4507-810c-ffdf3c0bce0d)

I am sorry if the text description in the vignette wasn't very clear. I’ve added these two GIFs to help clarify the process.
