### 2025-06-23 #7971 @ajstewartlang
> [@ajstewartlang](https://github.com/ajstewartlang) I've addressed all the comments from the reviewer and have not received any new ones. Can we move forward with the paper?

Thanks. We need to wait for the reviewers to take a look. @cmaimone - when you have a moment could you go through how your comments have been addressed please and share your views? @rhenkin and @kenbenoit - can I check how your reviews are coming along please?
