### 2025-06-23 chaoliu-cl/textAnnotatoR#11 @cmaimone
Interface issues/confusing things
Re: https://github.com/openjournals/joss-reviews/issues/7971

Sorry to put multiple items together - I'll leave it to you though to triage and separate out into multiple issues as needed.

1. I had one text file open and coded it.  I then opened a second text file.  It A) duplicated/repeated the text from that file multiple time in the display, and B) put empty code markers corresponding to what had been in the previous file showing at the end of the file:

2. Some of the code highlighting colors are too dark to read the highlighted text.  (See example below)

3. I added coded to a file and then went to the Themes tab.  I did not group codes into any themes.  I was still surprised though to see the stats box list Total Codes: 0 -- I'm assuming that means no codes associated with any themes, but I found it confusing to show 0, and for the hierarchy to not show codes that were not part of themes.  I would have expected them to be under Root by default.

4. I was surprised that Code co-occurrence was defined as the number of overlapping characters in highlights associated with different codes.  I've generally seen co-occurrence as codes in the same unit of text, with that unit sometimes being a sentence, paragraph, or file.

5. I do not see a column in the table in the records tab that records what file an annotation/highlight comes from.

6. I saved the project once, then made some changes, and then clicked save again (saving frequently as directed in the vignette).  I was surprised to have to choose the folder/location again to save the project file.  I understand that someone may want to put it in a different location, but I expected it to remember where it was saved.
