### 2025-06-25 fd79f7d
update package functions for JOSS
