+++ R/utils.R
@@ -43,6 +43,181 @@ handle_error <- function(expr, success_msg = NULL, error_msg = NULL, finally_msg
   })
 }
 
+#' Generate readable highlight color
+#'
+#' @description
+#' Generates a random color that provides good contrast with black text by ensuring
+#' the color is bright enough. Uses HSL color space to maintain good saturation
+#' while ensuring sufficient lightness.
+#'
+#' @return Character string containing a hex color code
+#' @keywords internal
+generate_readable_color <- function() {
+  # Generate random hue (0-360)
+  hue <- sample(0:359, 1)
+
+  # Set saturation between 40-70% for vibrant but not overwhelming colors
+  saturation <- sample(40:70, 1)
+
+  # Set lightness between 70-90% to ensure readability with black text
+  lightness <- sample(70:90, 1)
+
+  # Convert HSL to RGB
+  rgb_color <- hsl_to_rgb(hue, saturation, lightness)
+
+  # Return as hex color
+  return(sprintf("#%02X%02X%02X", rgb_color[1], rgb_color[2], rgb_color[3]))
+}
+
+#' Convert HSL to RGB color values
+#'
+#' @description
+#' Converts HSL (Hue, Saturation, Lightness) color values to RGB values.
+#' This allows for better control over color brightness and readability.
+#'
+#' @param h Numeric, hue value (0-360)
+#' @param s Numeric, saturation percentage (0-100)
+#' @param l Numeric, lightness percentage (0-100)
+#'
+#' @return Numeric vector of length 3 containing RGB values (0-255)
+#' @keywords internal
+hsl_to_rgb <- function(h, s, l) {
+  # Normalize values
+  h <- h / 360
+  s <- s / 100
+  l <- l / 100
+
+  if (s == 0) {
+    # Achromatic (grey)
+    r <- g <- b <- l
+  } else {
+    hue_to_rgb <- function(p, q, t) {
+      if (t < 0) t <- t + 1
+      if (t > 1) t <- t - 1
+      if (t < 1/6) return(p + (q - p) * 6 * t)
+      if (t < 1/2) return(q)
+      if (t < 2/3) return(p + (q - p) * (2/3 - t) * 6)
+      return(p)
+    }
+
+    q <- if (l < 0.5) l * (1 + s) else l + s - l * s
+    p <- 2 * l - q
+
+    r <- hue_to_rgb(p, q, h + 1/3)
+    g <- hue_to_rgb(p, q, h)
+    b <- hue_to_rgb(p, q, h - 1/3)
+  }
+
+  # Convert to 0-255 range and round
+  return(round(c(r, g, b) * 255))
+}
+
+#' Generate predefined readable color palette
+#'
+#' @description
+#' Provides a set of predefined colors that are known to have good readability
+#' with black text. This can be used as a fallback or alternative to random generation.
+#'
+#' @return Character vector of hex color codes
+#' @keywords internal
+get_readable_color_palette <- function() {
+  c(
+    "#FFE6CC", "#E6F3FF", "#E6FFE6", "#FFE6F3", "#F3E6FF", "#FFFFE6",
+    "#FFE6E6", "#E6F7FF", "#F0FFE6", "#FFE6DC", "#E6E6FF", "#E6FFDC",
+    "#FFCCCC", "#CCFFCC", "#CCCCFF", "#FFCCFF", "#CCFFFF", "#FFFFCC",
+    "#FFE0CC", "#E0FFCC", "#CCE0FF", "#FFCCE0", "#E0CCFF", "#E0FFE0",
+    "#FFEFD5", "#F0FFF0", "#F0F8FF", "#FFF0F5", "#F5F0FF", "#FFFACD",
[…]
