+++ man/create_empty_cooccurrence_result.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/annotate_gui.R
+\name{create_empty_cooccurrence_result}
+\alias{create_empty_cooccurrence_result}
+\title{Create empty co-occurrence result structure}
+\usage{
+create_empty_cooccurrence_result(unit)
+}
+\arguments{
+\item{unit}{Character string specifying the analytical unit}
+}
+\value{
+List with empty co-occurrence analysis results
+}
+\description{
+Returns an empty result structure when no valid data is available.
+}
+\keyword{internal}

[…]
