### 2025-06-25 d61daae
update package functions for JOSS
