+++ R/annotate_gui.R
@@ -54,7 +54,7 @@
 #' @importFrom shiny runApp shinyApp fluidPage actionButton observeEvent renderUI
 #'   showNotification showModal modalDialog removeModal updateTextAreaInput
 #'   updateTextInput tabPanel fileInput renderTable renderPlot plotOutput
-#'   tableOutput textInput textAreaInput selectInput checkboxGroupInput
+#'   tableOutput textInput textOutput checkboxInput textAreaInput selectInput checkboxGroupInput
 #' @importFrom shinyjs useShinyjs toggle runjs
 #' @importFrom data.tree Node
 #' @importFrom jsonlite fromJSON toJSON write_json
@@ -306,6 +306,17 @@ annotate_gui <- function() {
             width = NULL,
             id = "tabset",
             tabPanel("File", icon = icon("file"),
+                     # Add current file display
+                     conditionalPanel(
+                       condition = "output.current_file_name",
+                       div(style = "margin-bottom: 15px; padding: 10px; background-color: #e8f4e8; border-radius: 5px;",
+                           tags$p(
+                             icon("file", class = "text-success"),
+                             strong("Current file: "),
+                             textOutput("current_file_name", inline = TRUE)
+                           )
+                       )
+                     ),
                      fileInput("file_input", "Choose File",
                                accept = c(".txt", ".docx", ".pdf")),
                      actionButton("import_text", "Import Text")
@@ -354,10 +365,38 @@ annotate_gui <- function() {
             #         actionButton("generate_codebook", "Generate Code Book", icon = icon("book"))
             #),
             tabPanel("Analysis", icon = icon("chart-bar"),
-                     actionButton("code_frequency", "Code Frequency", icon = icon("chart-bar")),
-                     actionButton("code_co_occurrence", "Code Co-occurrence", icon = icon("project-diagram")),
-                     actionButton("word_cloud", "Word Cloud", icon = icon("cloud")),
-                     actionButton("text_summary", "Text Summary", icon = icon("file-alt"))
+                     fluidRow(
+                       column(6,
+                              wellPanel(
+                                h4("Analysis Settings"),
+                                selectInput("cooccurrence_unit",
+                                            "Co-occurrence Analysis Unit:",
+                                            choices = c("Paragraph" = "paragraph",
+                                                        "Sentence" = "sentence",
+                                                        "Document" = "document"),
+                                            selected = "paragraph"),
+                                helpText("Select the analytical unit for co-occurrence analysis:",
+                                         tags$br(),
+                                         tags$strong("Sentence:"), " Codes appearing in the same sentence",
+                                         tags$br(),
+                                         tags$strong("Paragraph:"), " Codes appearing in the same paragraph",
+                                         tags$br(),
+                                         tags$strong("Document:"), " Codes appearing anywhere in the document")
+                              )
+                       ),
+                       column(6,
+                              wellPanel(
+                                h4("Analysis Tools"),
+                                actionButton("code_frequency", "Code Frequency", icon = icon("chart-bar")),
+                                br(), br(),
+                                actionButton("code_co_occurrence", "Code Co-occurrence", icon = icon("project-diagram")),
+                                br(), br(),
+                                actionButton("word_cloud", "Word Cloud", icon = icon("cloud")),
+                                br(), br(),
[…]
