### 2025-07-06 2e8c482
update the paper and the package
