+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'textAnnotatoR: An Interactive Text Annotation Tool with ''shiny'' GUI'
+title: 'textAnnotatoR: An Interactive Text Annotation Tool with 'shiny' GUI'
 tags:
   - R
   - text analysis
@@ -19,115 +19,179 @@ bibliography: paper.bib
 
 # Summary
 
- textAnnotatoR is an interactive text annotation tool implemented in R using the Shiny framework. It provides researchers with a user-friendly graphical interface for qualitative text analysis that combines systematic coding, annotation, and analysis of textual data. The tool addresses the need for accessible and open-source solutions in qualitative research, particularly for researchers who work within the R ecosystem but require the functionality typically found in commercial qualitative data analysis software.
-
-The package facilitates the entire workflow of qualitative text analysis, from importing text documents to generating analytical insights. Researchers can highlight text segments, apply codes, create memos, organize codes hierarchically into themes, and analyze coding patterns. The interactive interface makes the coding process intuitive while maintaining the rigor required for qualitative research.
+textAnnotatoR is an interactive text annotation tool implemented in R using the Shiny framework. It provides researchers with a user-friendly graphical interface for qualitative text analysis that combines systematic coding, annotation, and analysis of textual data. The tool addresses the need for accessible, open-source solutions in qualitative research, particularly for researchers conducting interview analysis, focus group studies, literature reviews, and content analysis who require seamless integration with R's statistical and visualization capabilities. The package facilitates the entire workflow of qualitative text analysis, from importing text documents to generating analytical insights through an intuitive interface that maintains the rigor required for qualitative research.
 
 # Statement of need
 
-Qualitative text analysis is a key research methodology in social sciences, humanities, education, and other fields requiring in-depth understanding of textual data. While commercial software packages such as ATLAS.ti, MAXQDA, and NVivo offer robust functionality for qualitative data analysis (QDA), they are often expensive and offer limited compatibility with other research tools [hart2017qualitative]. The discontinuation of RQDA, a once-popular open-source QDA software [huang2014rqda], has further created a gap for researchers seeking cost-effective tools for coding and analyzing qualitative data.
+Qualitative text analysis is fundamental to research in social sciences, humanities, education, and other fields requiring in-depth understanding of textual data. Researchers typically face a choice between expensive commercial software and limited open-source alternatives, each with significant trade-offs.
 
-## Positioning textAnnotatoR in the Landscape of Qualitative Analysis Tools
+## Commercial Tools and Their Limitations
 
-The qualitative research software landscape includes several open-source alternatives such as Taguette, LaMa, qc, and corporaexplorer, each with distinct strengths and limitations:
+Leading commercial qualitative data analysis (QDA) software packages including ATLAS.ti, MAXQDA, and NVivo offer comprehensive functionality but present several barriers:
 
-- **Taguette** [Rampin2021] offers a user-friendly interface for text annotation with hierarchical coding but lacks integration with the R ecosystem and advanced analytical capabilities.
+- **Cost barriers**: License fees, which can range from hundreds to thousands of dollars annually, pose a significant financial barrier for individual researchers, students, and budget-constrained institutions
+- **Workflow integration**: These standalone applications operate in isolation from statistical analysis environments that require researchers to export data for quantitative analysis or visualization in separate tools
+ - **Reproducibility challenges**: Commercial tools often use proprietary formats and closed-source algorithms, making it difficult to document and reproduce analytical workflows
+- **Learning curve**: Each tool requires mastering a unique interface and workflow, with limited transferability between platforms
 
[…]
