+++ R/project.R
@@ -161,7 +161,7 @@ load_project_state <- function(filename) {
 #' @keywords internal
 save_as_html <- function(filename, rv) {
   # Get the current state of the text display
-  html_content <- update_text_display()
+  html_content <- update_text_display(rv)
 
   # Create a complete HTML document
   full_html <- paste0(
@@ -203,199 +203,96 @@ save_as_text <- function(filename, rv) {
   writeLines(annotated_text, filename)
 }
 
-#' Create plain text version of annotations
+#' Enhanced save project interactive dialog
 #'
 #' @description
-#' Converts annotated text to plain text format with code markers. Each annotation
-#' is represented as a code identifier and annotated text wrapped in square brackets.
-#' Multiple annotations are preserved and shown in order of appearance in the text.
-#'
-#' @param text Character string containing the original text
-#' @param annotations Data frame of annotations with columns:
-#'   \itemize{
-#'     \item start: Numeric vector of starting positions
-#'     \item end: Numeric vector of ending positions
-#'     \item code: Character vector of code names
-#'   }
-#'
-#' @return Character string containing formatted text with code markers
-#' @keywords internal
-create_plain_text_annotations <- function(text, annotations) {
-  if (nrow(annotations) == 0) {
-    return(text)
-  }
-
-  sorted_annotations <- annotations[order(annotations$start), ]
-  plain_text <- ""
-  last_end <- 0
-
-  for (i in 1:nrow(sorted_annotations)) {
-    if (sorted_annotations$start[i] > last_end + 1) {
-      plain_text <- paste0(plain_text, substr(text, last_end + 1, sorted_annotations$start[i] - 1))
-    }
-
-    # Include source file information in the annotation marker
-    source_info <- if ("source_file" %in% colnames(sorted_annotations) &&
-                       !is.na(sorted_annotations$source_file[i]) &&
-                       sorted_annotations$source_file[i] != "") {
-      paste0(" (", sorted_annotations$source_file[i], ")")
-    } else {
-      ""
-    }
-
-    plain_text <- paste0(plain_text,
-                         "[", sorted_annotations$code[i], source_info, ": ",
-                         substr(text, sorted_annotations$start[i], sorted_annotations$end[i]),
-                         "]")
-    last_end <- sorted_annotations$end[i]
-  }
-
-  if (last_end < nchar(text)) {
-    plain_text <- paste0(plain_text, substr(text, last_end + 1, nchar(text)))
-  }
-
-  return(plain_text)
-}
-
-# Update the apply_action function to handle source_file
-apply_action <- function(rv, action, reverse = FALSE) {
-  data <- if (reverse) action$reverse_data else action$data
-
-  switch(action$type,
-         "add_annotation" = {
-           if (reverse) {
-             # Remove annotation
-             if(nrow(rv$annotations) > 0) {
-               rv$annotations <- rv$annotations[-which(
-                 rv$annotations$start == data$start &
-                   rv$annotations$end == data$end &
-                   rv$annotations$code == data$code
-               ), ]
-             }
-           } else {
-             # Add annotation
-             if(is.null(rv$annotations)) {
-               rv$annotations <- data.frame(
-                 start = integer(),
-                 end = integer(),
-                 text = character(),
-                 code = character(),
-                 memo = character(),
-                 source_file = character(),
-                 stringsAsFactors = FALSE
-               )
-             }
-
-             # Ensure data has source_file column
-             if (!"source_file" %in% colnames(data)) {
-               data$source_file <- rv$current_file_name %||% "Unknown"
-             }
-
-             rv$annotations <- rbind(rv$annotations, data)
-           }
-         },
-         "merge_codes" = {
[…]
