+++ man/is_color_readable.Rd
@@ -1,15 +1,13 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
+% Please edit documentation in R/annotate_gui.R
 \name{is_color_readable}
 \alias{is_color_readable}
 \title{Validate color readability}
 \usage{
-is_color_readable(color, min_luminance = 0.6)
+is_color_readable(color)
 }
 \arguments{
 \item{color}{Character string containing hex color code}
-
-\item{min_luminance}{Numeric, minimum luminance threshold (0-1)}
 }
 \value{
 Logical indicating whether color is readable

[…]
