### 2025-07-06 4f2f0a7
Deploying to gh-pages from @ chaoliu-cl/textAnnotatoR@2e8c4826bd69ce3e187ac19996f5d416e52d6025 🚀
