### 2025-07-06 d7a24cc
update the paper and the package
