+++ man/parse_document.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/annotate_gui.R
-\name{parse_document}
-\alias{parse_document}
-\title{Parse text as single document unit}
-\usage{
-parse_document(text)
-}
-\arguments{
-\item{text}{Character string to parse}
-}
-\value{
-Data frame with single document unit
-}
-\description{
-Treats the entire text as a single analytical unit.
-}
-\keyword{internal}
