+++ man/calculate_unit_cooccurrence_matrix.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/annotate_gui.R
-\name{calculate_unit_cooccurrence_matrix}
-\alias{calculate_unit_cooccurrence_matrix}
-\title{Calculate unit-based co-occurrence matrix}
-\usage{
-calculate_unit_cooccurrence_matrix(codes, unit_assignments)
-}
-\arguments{
-\item{codes}{Character vector of unique codes}
-
-\item{unit_assignments}{Data frame from assign_annotations_to_units}
-}
-\value{
-Symmetric matrix with co-occurrence counts
-}
-\description{
-Creates a matrix showing how often pairs of codes appear together in the same analytical units.
-}
-\keyword{internal}
