+++ man/get_readable_color_palette.Rd
@@ -1,16 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
-\name{get_readable_color_palette}
-\alias{get_readable_color_palette}
-\title{Generate predefined readable color palette}
-\usage{
-get_readable_color_palette()
-}
-\value{
-Character vector of hex color codes
-}
-\description{
-Provides a set of predefined colors that are known to have good readability
-with black text. This can be used as a fallback or alternative to random generation.
-}
-\keyword{internal}
