+++ man/parse_paragraphs.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/annotate_gui.R
-\name{parse_paragraphs}
-\alias{parse_paragraphs}
-\title{Parse text into paragraphs}
-\usage{
-parse_paragraphs(text)
-}
-\arguments{
-\item{text}{Character string to parse}
-}
-\value{
-Data frame of paragraph units
-}
-\description{
-Splits text into paragraph units using line breaks.
-}
-\keyword{internal}
