+++ man/calculate_jaccard_similarity_matrix.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/annotate_gui.R
-\name{calculate_jaccard_similarity_matrix}
-\alias{calculate_jaccard_similarity_matrix}
-\title{Calculate Jaccard similarity matrix for unit-based co-occurrence}
-\usage{
-calculate_jaccard_similarity_matrix(codes, unit_assignments)
-}
-\arguments{
-\item{codes}{Character vector of unique codes}
-
-\item{unit_assignments}{Data frame from assign_annotations_to_units}
-}
-\value{
-Symmetric matrix with Jaccard similarity coefficients
-}
-\description{
-Calculates Jaccard similarity coefficients based on unit co-occurrence patterns.
-}
-\keyword{internal}
