+++ man/compare_unit_usage.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/annotate_gui.R
-\name{compare_unit_usage}
-\alias{compare_unit_usage}
-\title{Compare unit usage patterns between coders}
-\usage{
-compare_unit_usage(coding_strategies)
-}
-\arguments{
-\item{coding_strategies}{List of coding strategies with unit information}
-}
-\value{
-List containing unit usage comparison metrics
-}
-\description{
-Analyzes how different coders utilize analytical units in their coding patterns.
-}
-\keyword{internal}
