+++ man/parse_text_into_units.Rd
@@ -1,26 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/annotate_gui.R
-\name{parse_text_into_units}
-\alias{parse_text_into_units}
-\title{Parse text into analytical units}
-\usage{
-parse_text_into_units(text, unit)
-}
-\arguments{
-\item{text}{Character string containing the text to parse}
-
-\item{unit}{Character string specifying unit type: "sentence", "paragraph", or "document"}
-}
-\value{
-Data frame with columns:
-\itemize{
-\item unit_id: unique identifier for each unit
-\item start: starting character position of unit
-\item end: ending character position of unit
-\item text: text content of the unit
-}
-}
-\description{
-Divides text into sentences, paragraphs, or document-level units for co-occurrence analysis.
-}
-\keyword{internal}
