+++ man/calculate_phi_coefficient_matrix.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/annotate_gui.R
-\name{calculate_phi_coefficient_matrix}
-\alias{calculate_phi_coefficient_matrix}
-\title{Calculate Phi coefficient matrix for unit-based co-occurrence}
-\usage{
-calculate_phi_coefficient_matrix(codes, unit_assignments)
-}
-\arguments{
-\item{codes}{Character vector of unique codes}
-
-\item{unit_assignments}{Data frame from assign_annotations_to_units}
-}
-\value{
-Symmetric matrix with Phi coefficients
-}
-\description{
-Calculates Phi coefficients based on unit co-occurrence patterns.
-}
-\keyword{internal}
