+++ man/calculate_cooccurrence_summary.Rd
@@ -1,32 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/annotate_gui.R
-\name{calculate_cooccurrence_summary}
-\alias{calculate_cooccurrence_summary}
-\title{Calculate summary statistics for co-occurrence analysis}
-\usage{
-calculate_cooccurrence_summary(co_matrix, jaccard_matrix, phi_matrix, codes)
-}
-\arguments{
-\item{co_matrix}{Matrix of raw co-occurrence counts}
-
-\item{jaccard_matrix}{Matrix of Jaccard similarity coefficients}
-
-\item{phi_matrix}{Matrix of Phi coefficients}
-
-\item{codes}{Character vector of code names}
-}
-\value{
-List containing summary statistics:
-\itemize{
-\item total_codes: Total number of codes in analysis
-\item max_co_occurrence: Maximum co-occurrence count
-\item max_jaccard: Maximum Jaccard similarity coefficient
-\item mean_jaccard: Mean Jaccard similarity coefficient (excluding diagonal)
-#' \item significant_pairs: Number of code pairs with |phi| > 0.3
-}
-}
-\description{
-Calculates summary statistics for code co-occurrence matrices including
-maximum values, means, and counts of significant relationships.
-}
-\keyword{internal}
