+++ pkgdown.yml
@@ -4,4 +4,4 @@ pkgdown_sha: ~
 articles:
   practical_example: practical_example.html
   technical_integration: technical_integration.html
-last_built: 2025-07-06T02:37Z
+last_built: 2025-07-06T03:11Z
