### 2025-07-06 #7971 @chaoliu-cl
Hi @rhenkin,

Thank you again for your thoughtful and constructive feedback. I’ve made several major improvements to both the paper and the software in response to your suggestions.

**Revisions to the Paper**
The revised manuscript now:

1. Provides a more comprehensive overview of open-source QDA tools, including Taguette
2. Strengthens the case for R integration with detailed comparative analysis
3. Removes redundant content across sections
4. Incorporates use cases directly into the introduction
5. Replaces abstract examples with detailed, step-by-step visual demonstrations
6. Streamlines tool comparisons to avoid repetition

**Enhancements to the Software**

- **Annotation vs. Code Management**
  - Added a `remove_annotation()` function to delete specific annotations without affecting the underlying code
  - Redesigned the right-click popup menu with clearer separation of annotation-level and code-level actions:
     - Remove Annotation, Replace Code
     - — (separator) —
     - Rename Code, Delete Code

- **Dedicated Codebook Tab**
    - Centralized interface for code creation, renaming, merging, and deletion
    - Displays usage statistics and metadata for easier code management

- **Improved Labels and Tooltips**
    - Renamed the “Save” button in code creation to “**Create Code**” with a plus icon for clarity
    - Added descriptive Bootstrap tooltips to all icon-only buttons (e.g., Select, Clear, Undo, Redo), with proper positioning and dynamic updates

- **Performance Optimization**
    - Optimized importing and annotation functions to reduce unnecessary DOM updates and improve responsiveness by minimizing character-level processing

Please install the latest version of the package from GitHub:
`devtools::install_github("chaoliu-cl/textAnnotatoR")`
