### 2025-07-06 dcce8c3
Update paper.md
