### 2025-07-11 cda36f4
Codebook details and UI update
