+++ R/annotate_gui.R
@@ -84,7 +84,7 @@ annotate_gui <- function() {
   })
 
   ui <- dashboardPage(
-    dashboardHeader(title = span("textAnnotatoR", class = "brand-text")),
+    dashboardHeader(title = "textAnnotatoR", titleWidth = 250),
     dashboardSidebar(disable = TRUE),
     dashboardBody(
       useShinyjs(),
@@ -93,11 +93,15 @@ annotate_gui <- function() {
                div(style = "margin-bottom: 15px",
                    actionButton("save_project", "Save Project", icon = icon("save")),
                    actionButton("load_project", "Load Project", icon = icon("folder-open")),
-                   actionButton("new_project", "New Project", icon = icon("file"))
+                   actionButton("new_project", "New Project", icon = icon("file")),
+                   div(style = "display: inline-block; margin-left: 15px; border-left: 1px solid #ddd; padding-left: 15px;",
+                       actionButton("export_hierarchy_toolbar", "Export Hierarchy", icon = icon("download")),
+                       actionButton("import_hierarchy_toolbar", "Import Hierarchy", icon = icon("upload"))
+                   )
                )
         ),
         column(4,
-               div(id = "project_status", 
+               div(id = "project_status",
                    style = "margin-bottom: 15px; text-align: right; padding-top: 7px;",
                    uiOutput("project_status_display")
                )
@@ -229,6 +233,14 @@ annotate_gui <- function() {
             height: calc(100% - 40px);
             overflow-y: auto;
           }
+          .main-header .logo {
+            font-family: 'Gamja Flower', sans-serif !important;
+            font-weight: 800 !important;
+            letter-spacing: 1px !important;
+            text-transform: none !important;
+            color: #FFFFFF !important;
+            font-size: 24px !important;
+          }
           .theme-item, .code-item {
             display: inline-block;
             padding: 2px 4px;
@@ -259,24 +271,24 @@ annotate_gui <- function() {
           font-family: monospace;
           line-height: 1.5;
         }
-        
+
         /* Selected item styling */
         .theme-item.selected, .code-item.selected {
           background-color: #e6f3ff;
           border-radius: 3px;
           padding: 2px 4px;
         }
-        
+
         /* Theme item styling */
         .theme-item {
           font-weight: bold;
         }
-        
+
         /* Code item styling */
         .code-item {
           font-style: normal;
         }
-        
+
         /* Description preview styling */
         .description-preview {
           color: #666;
@@ -298,23 +310,23 @@ annotate_gui <- function() {
         $(document).ready(function() {
         // Initialize onclick handlers for themes and codes
         initializeThemeHandlers();
-        
+
         // Function to set up the click handlers
         function initializeThemeHandlers() {
           // Add click handler for theme and code items
           $(document).on('click', '.theme-item, .code-item', function(e) {
             e.preventDefault();
             e.stopPropagation();
-      
+
             // Remove selection from all items
             $('.theme-item, .code-item').removeClass('selected');
-            
+
             // Add selection to clicked item
             $(this).addClass('selected');
-      
+
             // Get the item name - this is the critical part
             var itemName = $(this).data('name');
-            
+
             // Make sure we have a valid name
             if (!itemName) {
               itemName = $(this).text().trim();
@@ -323,11 +335,11 @@ annotate_gui <- function() {
                 itemName = itemName.split(' - ')[0].trim();
               }
             }
-      
+
             // Send the selected item to Shiny
             Shiny.setInputValue('selected_theme', itemName);
           });
-      
+
           // Add hover effects for better UX
           $(document).on('mouseenter', '.theme-item, .code-item', function() {
             $(this).css({
@@ -358,14 +370,14 @@ annotate_gui <- function() {
[…]
