+++ man/get_closed_folder_icon.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils.R
+\name{get_closed_folder_icon}
+\alias{get_closed_folder_icon}
+\title{Get HTML entity for closed folder icon}
+\usage{
+get_closed_folder_icon()
+}
+\value{
+Character string containing HTML entity
+}
+\description{
+Returns the HTML entity code for a closed folder icon
+}
+\keyword{internal}
