+++ DESCRIPTION
@@ -4,7 +4,7 @@ Title: Interactive Text Annotation Tool with 'shiny' GUI
 Version: 0.1.4
 Date: 2024-10-25
 Authors@R: person("Chao Liu", email = "chaoliu@cedarville.edu", role = c("aut", "cre", "cph"), comment = c(ORCID = "0000-0002-9979-8272"))
-Description: A comprehensive text annotation tool built with 'shiny'. Provides an 
+Description: A A lightweight and focused text annotation tool built with 'shiny'. Provides an 
     interactive graphical user interface for coding text documents,
     managing code hierarchies, creating memos, and analyzing coding patterns.
     Features include code co-occurrence analysis, visualization of coding patterns,
