### 2025-07-11 5459548
Codebook details and UI update
