### 2025-07-11 #7971 @chaoliu-cl
Thank you @rhenkin again for your feedback.

The app has been updated once again to address all the issues you raised. Regarding point #4, I removed the colored buttons for “Create Code” and “Merge Selected” to align with Bootstrap’s default button styling, which offers a cleaner and more consistent appearance. I’ve kept the color for “Delete Selected,” as it appropriately signals a destructive action.

As you suggested, I’ve also added the Export Hierarchy and Import Hierarchy buttons to the main toolbar for easier access.

That said, I want to clarify that textAnnotatoR, as the name implies, is intentionally designed with a focused scope: to support core annotation functionality rather than serve as a full-scale project management system. The goal is to provide a clean, intuitive interface that remains accessible to non-technical researchers.

Managing multiple large text files simultaneously within a Shiny app may introduce performance challenges, particularly for users with limited computing resources. With that in mind, I would prefer for _textAnnotatoR_ to prioritize robustness and maintainability over expanding its feature set, in keeping with the principle of “doing one thing well.”
