### 2025-07-14 #7971 @rhenkin
@ajstewartlang I'm done with my review, I think the tool is in a publishable state and I could use the interface as intended and I believe he current workflow works much better. I do note the limitation discussed above regarding the number of files that can be opened at once. I think it should not be an obstacle to publication but I think the paper and other software documentation should make it very clear.

@chaoliu-cl I think it would be fair to state this limitation in the paper.
