### 2025-07-15 #7971 @ajstewartlang
Many thanks @rhenkin for your review. @chaoliu-cl can you add that limitation please and let me know when it's done please @kenbenoit, have you had a chance to look at this yet?
