### 2025-07-16 3a5c819
update JOSS paper
