### 2025-07-16 #7971 @chaoliu-cl
@ajstewartlang I've updated the paper again to include a clarification of the program's focused scope and to add a note about its design constraints near the end of the Implementation section.
