### 2025-08-01 #7971 @kbenoit
I like the objective of this package and think that the implementation is a nice balance between simplicity and functionality. I can see how this could be useful for many needing to perform annotation on a budget. Is it really a better solution than NVivo or QDAMiner? I'm not sure, although the price is certainly a winner! I wonder if the reproducibility claims really apply to this package as an advantage over those products.

Nonetheless I think it's a valuable contribution to the text analysis ecosystem in R and think it should be accepted.

A few minor issues first though.

#### Installation: Does installation proceed as outlined in the documentation?

- I had some trouble with my local build but fixed them in https://github.com/chaoliu-cl/textAnnotatoR/pull/18, https://github.com/chaoliu-cl/textAnnotatoR/pull/17, and https://github.com/chaoliu-cl/textAnnotatoR/pull/16.

#### Functionality; Performance

In the texts I tried, it was finicky to get the text to be selected. Sometimes it highlighted them in yellow, other times, I could not get the yellow to come on.  Before reverting to the CRAN version, I could not get the highlighting/selection to work at all.

These claims have not been fully confirmed in terms of "Reproducible research: All analytical steps can be documented in R scripts, enabling complete reproducibility of both qualitative coding decisions and subsequent analyses—a critical requirement for transparent and rigorous research." How is the output more reproducible than other annotation packages?

#### Automated tests

- Yes, there are tests, but it is not easy to test the functionality of the interactive parts of the Shiny app.
- suggest adding a code coverage badge to `README.Rmd`

#### Installation instructions

Yes these are clear, but I tidied up the Imports in https://github.com/chaoliu-cl/textAnnotatoR/pull/17.  Some functions had the wrong packages for importing, and there were packages in Imports: that didn't need to be declared.

#### Community guidelines

It would be a simple matter to add these using `usethis::use_code_of_conduct()`.

#### Misc article issues

In the article, the image should be resized to a smaller size on p9.
