### 2025-08-04 49b49cb
Text highlighting improvement and paper update
