+++ R/annotate_gui.R
@@ -129,14 +129,54 @@ annotate_gui <- function() {
 
           body { background-color: #f0f0f0; }
           .char { cursor: pointer; }
-          .highlighted { background-color: yellow; }
+          .highlighted {
+            background-color: #ffff00 !important;
+            box-shadow: 0 0 3px rgba(255, 255, 0, 0.8);
+          }
+
+          /* NEW: Selection mode styles */
+          .selection-active { cursor: crosshair !important; }
+          .selection-active .char { cursor: crosshair !important; }
+          .selection-active .char:hover {
+            background-color: rgba(255, 255, 0, 0.3);
+          }
+          .selection-mode-indicator {
+            position: fixed;
+            top: 10px;
+            right: 10px;
+            background-color: #28a745;
+            color: white;
+            padding: 5px 10px;
+            border-radius: 4px;
+            font-size: 12px;
+            z-index: 1001;
+            display: none;
+          }
+          .selection-mode-indicator.active {
+            display: block;
+          }
+
+          /* Toggle button styles */
+          .btn-toggle-active {
+            background-color: #007bff !important;
+            color: white !important;
+            border-color: #0056b3 !important;
+            box-shadow: inset 0 3px 5px rgba(0,0,0,.125) !important;
+          }
+
+          .btn-toggle-inactive {
+            background-color: #f8f9fa !important;
+            color: #6c757d !important;
+            border-color: #dee2e6 !important;
+          }
+
           .code-display {
             padding: 2px 5px;
             margin-right: 5px;
             border-radius: 3px;
             font-weight: bold;
             color: black;
-            display: inline-block; /* Ensure proper display with line breaks */
+            display: inline-block;
           }
           .code-display br {
             display: block;        /* Make line breaks more prominent */
@@ -698,6 +738,19 @@ annotate_gui <- function() {
       if (is.null(rv$storage_mode)) {
         init_data_dir(session)
       }
+    })
+
+    # Initialize selection button state
+    observe({
+      # Initialize button as inactive on startup
+          runjs('
+        $(document).ready(function() {
+          if ($("#selection-indicator").length === 0) {
+            $("body").append("<div id=\\"selection-indicator\\" class=\\"selection-mode-indicator\\">Selection Mode is Inactive</div>");
+          }
+          $("#select").addClass("btn-toggle-inactive");
+        });
+      ')
     })
 
     # Handle storage confirmation - fixed by passing output parameter
@@ -3157,109 +3210,182 @@ annotate_gui <- function() {
 
     # Modify the JavaScript for text selection
     observeEvent(input$select, {
-      runjs("
-            var startChar = null;
-            var endChar = null;
-            var isSelecting = false;
-            var activeWindow = 'main'; // Track which window is active for selection
+      runjs('
+        // Global variables for selection state
+        var selectionState = {
+          isActive: false,
+          startChar: null,
+          endChar: null,
+          isSelecting: false
+        };
 
-            function initializeSelection() {
-              // Remove existing event listeners
-              $('#text_display, #floating_text_content').off('mousedown mousemove');
-              $(document).off('mouseup');
+        // Check current state and toggle
+        var currentlyActive = $("#text_display, #floating_text_content").hasClass("selection-active");
+        var newState = !currentlyActive;
 
-              // Add mousedown event listener for character selection
-              $('#text_display, #floating_text_content').on('mousedown', '.char', function(e) {
-                e.preventDefault();
-                if (!isSelecting) return;
+        // Initialize selection functionality
+        function initializeTextSelection() {
+          // Remove any existing event listeners to prevent duplicates
+          $(document).off("mousedown.textselection mousemove.textselection mouseup.textselection");
+          $("#text_display, #floating_text_content").off("mousedown.textselection mousemove.textselection");
 
[…]
