+++ README.Rmd
@@ -16,7 +16,7 @@ knitr::opts_chunk$set(
 # textAnnotatoR: Interactive Text Annotation Tool for R
 
 [![R-CMD-check](https://github.com/chaoliu-cl/textAnnotatoR/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/chaoliu-cl/textAnnotatoR/actions/workflows/R-CMD-check.yaml)
-[![Codecov test coverage](https://codecov.io/gh/chaoliu-cl/textAnnotatoR/branch/main/graph/badge.svg)](https://codecov.io/gh/chaoliu-cl/textAnnotatoR?branch=main)
+[![Codecov test coverage](https://codecov.io/gh/chaoliu-cl/textAnnotatoR/branch/main/graph/badge.svg?token=a8a91118-d5d9-4267-a0fa-077c0fc6e3f2)](https://codecov.io/gh/chaoliu-cl/textAnnotatoR?branch=main)
 [![CRAN
 status](https://www.r-pkg.org/badges/version/textAnnotatoR)](https://CRAN.R-project.org/package=textAnnotatoR)
 [![Lifecycle:
