### 2025-08-04 cdae993
Adding test coverage and rerun R CMD check
