### 2025-08-04 f624b01
Deploying to gh-pages from @ chaoliu-cl/textAnnotatoR@96f46a4cae773c4f75dd1165c732e65e91d541a5 🚀
