### 2025-08-25 #7971 @chaoliu-cl
Hi @ajstewartlang, since we haven’t received any additional comments from the reviewers, would it make sense to move forward with the paper? It has been dragging on for a while. Thank you!
