### 2025-08-27 #7971 @kbenoit
All of my concerns have been addressed, thanks @chaoliu-cl. Except:

Installing and running the program, and importing texts, I still have trouble highlighting text. Example (see the message lower right):
