### 2025-08-28 #7971 @chaoliu-cl
Ah, I see the issue. You’ll need to click with the cursor first to enable text selection:

![Image](https://github.com/user-attachments/assets/68592c4e-70ef-41a1-8952-9c707ad897d5)

I’ve updated the vignette to make this clearer. Text selection and highlighting are working as intended.
