### 2025-08-28 aca67c3
Update a vignette figure
