### 2025-08-28 fc5720e
Deploying to gh-pages from @ chaoliu-cl/textAnnotatoR@3397bac6390463fde6592a890ff29735573d788e 🚀
