### 2025-08-29 e6fd076
Update paper DOIs
