+++ paper/paper.bib
@@ -22,7 +22,8 @@
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
   year = {2024},
-  url = {https://www.R-project.org/}
+  url = {https://www.R-project.org/},
+  doi = {10.32614/r.manuals}
 }
 
 @Manual{shiny,
@@ -40,6 +41,7 @@
     year = {2021},
     note = {R package version 0.7.2},
     url = {https://CRAN.R-project.org/package=shinydashboard},
+    doi = {10.32614/cran.package.shinydashboard}
 }
 
 @Manual{datatree,
@@ -48,6 +50,7 @@
     year = {2023},
     note = {R package version 1.1.0},
     url = {https://CRAN.R-project.org/package=data.tree},
+    doi = {10.32614/cran.package.data.tree}
 }
 
 @Manual{DT,
@@ -56,6 +59,7 @@
     year = {2024},
     note = {R package version 0.3.3},
     url = {https://CRAN.R-project.org/package=DT},
+    doi = {10.32614/cran.package.dt}
 }
 
 @Article{jsonlite,
@@ -72,7 +76,8 @@
     year = {2024},
     note = {R package version 0.9.1},
     url = {https://CRAN.R-project.org/package=readtext},
-  }
+    doi = {10.32614/cran.package.readtext}
+}
 
 @article{Bogachenkova2023,
   author = {Bogachenkova, Anastasia and others},
