### 2025-08-29 #7971 @chaoliu-cl
@ajstewartlang I have added the missing dois to those specified references.
