### 2025-08-29 8a2af8f
Increment version number to 1.0.0
