+++ NEWS.md
@@ -1,3 +1,5 @@
+# textAnnotatoR 1.0.0
+
 # textAnnotatoR 0.1.1
 
 * Initial CRAN submission.
