+++ NEWS.md
@@ -1,4 +1,4 @@
-# textAnnotatoR 1.0.0
+# textAnnotatoR 1.0.1
 
 # textAnnotatoR 0.1.1
 
