### 2025-08-30 #7971 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
